MangaVision_API/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
ultralytics==8.3.32
manga-ocr==0.1.13
pydantic==2.9.2
Pillow==11.0.0
numpy==2.1.3
shapely==2.0.6
torch==2.5.0 
torchvision==0.20.0 
torchaudio==2.5.0

